+++ CHANGELOG.md
@@ -314,3 +314,4 @@ Repaired PumpUpdateLiteralsTestCase.  Now working as expected.
 * **2015-09-16** Begin examples/locations. Location enumeration supports unicode.  order_by no longer required in 
 make_get_query. Remove uri print from do_get. Update examples/locations buildings and cities examples.
 * **2015-09-17** Add examples/concepts.
+* **2015-09-18** Move create_enum to vivopump.py.  Update make_enum in examples.